Refs 17138. Spliting implementation from public APIs
