+++ src/cpp/dynamic-types/v1_3/AnnotationManager.hpp
